move extended target capture tasks
